Fix bug in remove_vertex where removing a list containing the vertex of largest index would not update property maps
I've stumbled on the edge case where vmax is to be removed as part of a list and I'm kinda convinced it should be like this... at least it #worksforme, or seems to. I haven't checked if the internal C++ functions have an issue with handling vmax being moved/shfited.

I suspect the previous form, however buggy, might have intended not to waste resources updating property maps if you're removing vmax alone. But since I'm not sure of this intention, or whether it ultimately makes sense, I'll leave it as an exercise. ;)

Ni!